feat: add new api version v2
Version 2 introduces enchancements that enable website routers for single page applications.
