+++ packages/messenger-internal/src/shared/util.ts
