+++ packages/messenger-internal/src/v2/index.ts
@@ -1,7 +1,6 @@
 export * from "./types";
 
 export {
-  CoreApi,
   Credentials,
   MountOptions,
   State,
