+++ packages/messenger/src/index.ts
@@ -6,6 +6,7 @@ import {
   MessengerInfo,
   v0,
   v1,
+  v2,
 } from "@userlike/messenger-internal";
 
 export * from "@userlike/messenger-internal";
@@ -15,7 +16,16 @@ export * from "@userlike/messenger-internal";
  * version information and relevant dates.
  */
 export async function createMessenger(
-  opts: MessengerOptions<1>
+  opts: MessengerOptions<2>,
+): Promise<ActionResult<string, MessengerInfo<2, v2.Messenger>>>;
+
+/**
+ * For internal use only.
+ *
+ * @deprecated
+ */
+export async function createMessenger(
+  opts: MessengerOptions<1>,
 ): Promise<ActionResult<string, MessengerInfo<1, v1.Api>>>;
 
 /**
@@ -24,20 +34,20 @@ export async function createMessenger(
  * @deprecated
  */
 export async function createMessenger(
-  opts: MessengerOptions<0>
+  opts: MessengerOptions<0>,
 ): Promise<ActionResult<string, MessengerInfo<0, v0.Api>>>;
 
 export async function createMessenger(
-  opts: MessengerOptions<number>
+  opts: MessengerOptions<number>,
 ): Promise<ActionResult<string, MessengerInfo<number, VersionedApi<number>>>>;
 
 export async function createMessenger(
-  opts: MessengerOptions<number>
+  opts: MessengerOptions<number>,
 ): Promise<ActionResult<string, MessengerInfo<number, VersionedApi<number>>>> {
   const { createMessenger } = await loadScripts(
     window,
     opts.widgetKey,
-    opts.baseUrl
+    opts.baseUrl,
   );
 
   return createMessenger(opts.version)({
